+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [1.5.6]
+- **PiP Close Button**: Added a close button (×) in the top-right corner of Picture-in-Picture mode
 - **Status Bar**: Added toggle to show/hide the status bar. Added "Show Status Bar" item in the View menu (after "Show OSD Notifications").
 
 ## [1.5.5]